requirements.txt
# Core scientific stack
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
matplotlib==3.9.2
scikit-image==0.24.0
scikit-learn==1.5.1

# Geospatial stack
rasterio==1.3.10
shapely==2.0.5
folium==0.17.0

# SAR / ASF
asf-search==7.1.3

# Machine Learning (for scoring)
xgboost==2.1.1

# Utils
tqdm==4.66.5
Pillow==10.4.0